refactor: RepositoryAdapter -> RepositoryImpl 네이밍 변경 및 파일 위치 수정
